requirements.txt
Flask
python-dotenv
Flask-Limiter